refactor(sdk): extract streaming engine into agentwire/sdk/ package
Phase 1 of docs/missions/agentwire-sdk-primitives.md. Pulls the SDK
plumbing duplicated between the Textual REPL and the workflow runner
into a single reusable package. Zero behavior change — all 1128 tests
pass; agentwire repl + workflow runs are byte-identical.

New package layout:
- agentwire/sdk/client.py      — build_options + thinking_config + MCP +
                                 default model/effort/tools constants
- agentwire/sdk/render.py      — render_message + format helpers +
                                 heartbeat_iter + ANSI styling
- agentwire/sdk/state.py       — StreamRenderState (partial-message
                                 streaming state machine)
- agentwire/sdk/events.py      — SDK Message → pi-shape JSONL translation
                                 (relocated from workflows/runners/)
- agentwire/sdk/errors.py      — classify() error tagger (relocated)
- agentwire/sdk/capabilities.py — model/effort/thinking validator
                                 (relocated)
- agentwire/sdk/damage_control.py — in-process pre-tool damage check
                                   (relocated from repl/)
- agentwire/sdk/sinks/base.py  — Sink protocol (write/flush/isatty)
- agentwire/sdk/sinks/textual.py — RichLogSink, ActionSink (relocated
                                  from repl/textual_app.py)

Consumers refactored to import from agentwire.sdk:
- agentwire/repl/app.py shrank from 947 LOC to 169 LOC; print mode +
  transcript persistence are the only REPL-specific concerns left
- agentwire/repl/textual_app.py drops its inline _RichLogSink /
  _ActionSink / _StreamRenderState — uses sdk versions
- agentwire/workflows/runners/anthropic.py imports events/capabilities/
  errors from sdk (no logic change)
- agentwire/workflows/node.py imports validate_node_settings from sdk

This phase builds the engine. Phase 2 ships the fan-out N-column view
that proves the primitives are reusable; Phase 3 wires a websocket sink
for portal watch-mode.

Built by dotdev.dev
