+++ agentwire/sdk/capabilities.py
