+++ agentwire/workflows/runners/anthropic.py
@@ -25,9 +25,9 @@ from pathlib import Path
 from typing import Any, Callable
 
 from agentwire.workflows.node import ActionNode, NodeResult
-from agentwire.workflows.runners import anthropic_events as ev
-from agentwire.workflows.runners.anthropic_capabilities import validate_node_settings
-from agentwire.workflows.runners.sdk_errors import classify as _classify
+from agentwire.sdk import events as ev
+from agentwire.sdk.capabilities import validate_node_settings
+from agentwire.sdk.errors import classify as _classify
 
 logger = logging.getLogger("agentwire.workflows.anthropic")
